teach-mode/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
av==15.0.0
certifi==2025.6.15
charset-normalizer==3.4.2
coloredlogs==15.0.1
ctranslate2==4.6.0
distro==1.9.0
exceptiongroup==1.3.0
faster-whisper==1.1.1
filelock==3.18.0
flatbuffers==25.2.10
fsspec==2025.5.1
groq==0.29.0
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.2
humanfriendly==10.0
idna==3.10
keyboard==0.13.5
MouseInfo==0.1.3
mpmath==1.3.0
numpy
onnxruntime==1.22.0
packaging==25.0
pillow==11.3.0
protobuf==6.31.1
PyAutoGUI==0.9.54
pydantic==2.11.7
pydantic_core==2.33.2
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynput==1.8.1
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dotenv==1.1.1
pytweening==1.2.0
PyYAML==6.0.2
requests==2.32.4
rubicon-objc==0.5.1
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
tokenizers==0.21.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
